Extrai registro de módulos do compilador para arquivos separados
Refatora os métodos registrarModulo* de compilador-llvm.ts para
fontes/registro-modulos/ usando mixins, reduzindo o tamanho do
arquivo principal. Move a interface EntradaFuncaoModulo para seu
próprio arquivo e altera propriedades de private para protected.
